Исправил ошибки

+++ bankServer/src/main/java/com/example/bankserver/kafka/PersonsListeningService.java
@@ -1,8 +1,10 @@
 package com.example.bankserver.kafka;
 
+import org.jooq.impl.QOM;
+
 import java.util.List;
 
 public interface PersonsListeningService {
 
-    void listen(List<Long> personIds);
+    void listen(List<Integer> personIds);
 }
